requirements.txt
requests>=2.28.0
mcp>=1.0.0
firecrawl-py>=0.0.16